pyproject.toml
[project]
name = "qmanalysis"
version = "0.1.0"
description = ""
authors = [
    {name = "Markus Walther",email = "markus.walther@fau.de"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pyyaml (>=6.0,<7.0.0)",
    "numpy (>=2.3.1,<3.0.0)",
    "scipy (>=1.16.0,<2.0.0)",
    "pandas (>=2.3.0,<3.0.0)",
    "matplotlib (>=3.10.3,<4.0.0)",
    "asteval (>=1.0.6,<2.0.0)",
    "openpyxl (>=3.1.5,<4.0.0)"
]

[tool.poetry]
packages = [{include = "qmanalysis", from = "src"}]

[tool.poetry.scripts]
qmanalysis = "main:main"
docs-build = "scripts.docs_build:main"
makexyz = "tools.makexyz:main"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
pytest-cov = "^6.2.1"
sphinx-autodoc-typehints = "^3.2.0"
sphinx-rtd-theme = "^3.0.2"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
